pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "span-nilm"
version = "0.1.0"
description = "Device detection from SPAN smart panel data using NILM techniques"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "requests>=2.28",
    "pyyaml>=6.0",
    "pyarrow>=12.0",
]

[project.scripts]
span-nilm = "span_nilm.__main__:main"
